Agentic AI/knowledge-agent/requirements.txt
fastapi
uvicorn
duckduckgo-search
numpy
requests
trafilatura
sentence-transformers
langchain
langchain-groq
python-dotenv
